requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
webdriver-manager>=4.0.0
pytest>=7.4.0
pytest-mock>=3.12.0